[recnet-api] Deactivate account

+++ apps/recnet-api/src/modules/user/user.service.ts
@@ -88,6 +88,16 @@ export class UserService {
     return this.transformUser(user);
   }
 
+  public async updateUserActivate(
+    userId: string,
+    isActivated: boolean
+  ): Promise<User> {
+    const user: DbUser = await this.userRepository.updateUser(userId, {
+      isActivated,
+    });
+    return this.transformUser(user);
+  }
+
   public async validateHandle(handle: string): Promise<void> {
     const user = await this.userRepository.findUserByHandle(handle);
     if (user) {
@@ -118,9 +128,15 @@ export class UserService {
     userId: string,
     followingUserId: string
   ): Promise<void> {
-    // validate if the user exists
-    await this.userRepository.findUserById(userId);
-    await this.userRepository.findUserById(followingUserId);
+    // validate if the user exists and is activated
+    const followingUser =
+      await this.userRepository.findUserById(followingUserId);
+    if (!followingUser.isActivated) {
+      throw new RecnetError(
+        ErrorCode.ACCOUNT_NOT_ACTIVATED,
+        HttpStatus.BAD_REQUEST
+      );
+    }
 
     await this.followingRecordRepository.createFollowingRecord(
       userId,
@@ -132,9 +148,15 @@ export class UserService {
     userId: string,
     followingUserId: string
   ): Promise<void> {
-    // validate if the user exists
-    await this.userRepository.findUserById(userId);
-    await this.userRepository.findUserById(followingUserId);
+    // validate if the user exists and is activated
+    const followingUser =
+      await this.userRepository.findUserById(followingUserId);
+    if (!followingUser.isActivated) {
+      throw new RecnetError(
+        ErrorCode.ACCOUNT_NOT_ACTIVATED,
+        HttpStatus.BAD_REQUEST
+      );
+    }
 
     await this.followingRecordRepository.deleteFollowingRecord(
       userId,
@@ -166,6 +188,7 @@ export class UserService {
       numRecs: user.recommendations.length,
       email: user.email,
       role: user.role,
+      isActivated: user.isActivated,
       following: followingUsers,
     };
   }
